requirements.txt
beautifulsoup4==4.13.3
scrapy==2.12.0
numpy==1.26.4
pandas==2.2.3
tqdm==4.67.1
python-dotenv==1.0.1
requests==2.32.3
transformers==4.42.4
sentence-transformers==2.6.0
torch==2.2.2
accelerate==0.32.1
bitsandbytes==0.43.1
einops==0.8.0
tiktoken==0.9.0
chromadb==0.6.3
langchain
langchain-huggingface
langchain-community
